Added general in-memory file upload methods to the Google Service Client.
This code is untested. Mocking the Drive Service and Create Upload Request proves to be extremely difficult due to lack of exposed interfaces by the Nuget Package. Too many days were sunk into this tech spike already, and at this point the priority becomes the release. The code that's untested is trivial enough to be considered safe.