Speed up GenerateReportUseCase Tests (#150)
Speed up the HFS reports use case tests from ~15 minutes running time to ~1 minute by allowing file wait and retry durations to be modified.

This cuts down the CI pipeline duration to ~2.5 minutes, as the reports UC tests were the bottleneck. The reports UC made use of hard time waits and retry intervals which ran far longer than is needed in unit tests that aren't connected to real infrastructure.

This PR modifies the GenerateReportUseCase to take in optional sleepDuration and retryInterval params that retain the original values as defaults, but can be overridden in the tests or elsewhere in the application as needed. It also ensures the UC reports are generated and checked using a retry mechanism and a timeout similar to how it was already done in one of the reports, instead of just hard wait duration.

Time calculations in the tests often assumed the default values for sleep and retry durations, so they have been modified accordingly.