Define Serverless dependency versions
This change resolves a build failure because Serverless Framework 4 doesn't pick up globally installed modules. I'm not sure why this has changed (or how/if it's actually changed), but it's sensible to version the dependencies and installing the plugin using 'npm install' following the docs _also_ allows the command to succeed.